death_bot_panel/death_bot_panel/quotes/serializers.py
from rest_framework import serializers

from death_bot_panel.quotes.models import Quote


class QuoteResponseSerializer(serializers.ModelSerializer):

    class Meta:
        model = Quote
        exclude = []
