pyproject.toml
[tool.poetry]
name = "death-bot-panel"
version = "0.1.0"
description = "Death bot panel"
authors = ["a.s.lebedev <a.s.lebedev@fun-box.ru>"]
readme = "README.md"
packages = [{include = "death_bot_panel"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.6"
psycopg2 = "^2.9.5"
python-dotenv = "^0.21.1"
djangorestframework = "^3.14.0"


[tool.poetry.group.dev.dependencies]
black = "^23.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py310']
include = '\.pyi?$'
